+++ test/cpp/jit/no-gtest.cpp
@@ -1,4 +1,5 @@
-#include <test/cpp/jit/tests.h>
+#include <test/cpp/jit/test_alias_analysis.h>
+#include <test/cpp/jit/test_misc.h>
 
 #include <sstream>
 #include <string>
